+++ radio/RadioStation/NowPlayingResponse.swift
@@ -8,7 +8,7 @@ public struct NowPlayingResponse: Codable {
     let playingNext: PlayingNext
     let songHistory: [SongHistory]
     let isOnline: Bool
-    
+
     enum CodingKeys: String, CodingKey {
         case station
         case listeners
@@ -24,7 +24,7 @@ struct Station: Codable {
     let id: Int
     let name: String
     let listenURL: URL
-    
+
     enum CodingKeys: String, CodingKey {
         case id
         case name
@@ -36,7 +36,7 @@ struct Listeners: Codable {
     let total: Int
     let unique: Int
     let current: Int
-    
+
     enum CodingKeys: String, CodingKey {
         case total
         case unique
@@ -49,7 +49,7 @@ struct Live: Codable {
     let streamerName: String
     let broadcastStart: Date?
     let art: URL?
-    
+
     enum CodingKeys: String, CodingKey {
         case isLive = "is_live"
         case streamerName = "streamer_name"
@@ -69,7 +69,7 @@ struct Song: Codable {
     let lyrics: String
     let art: URL
     let customFields: [String]
-    
+
     enum CodingKeys: String, CodingKey {
         case id
         case text
@@ -87,14 +87,14 @@ struct Song: Codable {
 struct NowPlaying: Codable {
     let shID: Int
     let playedAt: TimeInterval
-    let duration: Int
+    let duration: Float
     let playlist: String
     let streamer: String
     let isRequest: Bool
     let song: Song
     let elapsed: Int
     let remaining: Int
-    
+
     enum CodingKeys: String, CodingKey {
         case shID = "sh_id"
         case playedAt = "played_at"
@@ -111,11 +111,11 @@ struct NowPlaying: Codable {
 struct PlayingNext: Codable {
     let cuedAt: TimeInterval
     let playedAt: TimeInterval
-    let duration: Int
+    let duration: Float
     let playlist: String
     let isRequest: Bool
     let song: Song
-    
+
     enum CodingKeys: String, CodingKey {
         case cuedAt = "cued_at"
         case playedAt = "played_at"
@@ -129,12 +129,12 @@ struct PlayingNext: Codable {
 struct SongHistory: Codable {
     let shID: Int
     let playedAt: TimeInterval
-    let duration: Int
+    let duration: Float
     let playlist: String
     let streamer: String
     let isRequest: Bool
     let song: Song
-    
+
     enum CodingKeys: String, CodingKey {
         case shID = "sh_id"
         case playedAt = "played_at"
